+++ tests/ota/files/external/urls.json
@@ -14,5 +14,9 @@
     "dl/local_provider/RDL2016091_1_E11-G13_V0.0.9_20170921_release.ota": {
         "url": "https://us-fm.cloud.sengled.com/sengled/zigbee/firmware/RDL2016091_1_E11-G13_V0.0.9_20170921_release.ota",
         "checksum": "sha3-256:542122a6f0f48075c3f089afcf9a4f86af2741672bc8c9cc7d5759ceed5c0e63"
+    },
+    "dl/sonoff/snzb-01m_v1.0.5.ota": {
+        "url": "https://raw.githubusercontent.com/Koenkk/zigbee-OTA/fb19e4fa0fc278135092a43f01d6f9044c5e65ee/images/Sonoff/snzb-01m_v1.0.5.ota",
+        "checksum": "sha3-256:5f50dcdd5c76dfcdf9c5166a3f3131f9b6164d0430914bb6b778d987022f8030"
     }
 }
